--- packages/venia-ui/lib/components/CreateAccountPage/helpers.js
@@ -1,10 +1,0 @@
-const validCreateAccountParams = ['email', 'firstName', 'lastName'];
-
-export const getCreateAccountInitialValues = search => {
-    const params = new URLSearchParams(search);
-
-    return validCreateAccountParams.reduce(
-        (values, param) => ({ ...values, [param]: params.get(param) }),
-        {}
-    );
-};
